Cardano.Logging.Configuration

Documentation

data ConfigReflection

Constructors

ConfigReflection

Fields

crSilent :: IORef (Set [Text])

crNoMetrics :: IORef (Set [Text])

crAllTracers :: IORef (Set [Text])

data HermodException

Constructors

HermodConfigException

Fields

excMessage :: String

emptyConfigReflection :: IO ConfigReflection

configureTracers :: forall a m. (MetaTrace a, MonadIO m) => ConfigReflection -> TraceConfig -> [Trace m a] -> m ()

Call this function at initialisation, and later for reconfiguration. Config reflection is used to optimise the tracers and has to collect information about the tracers. Although it is possible to give more then one tracer of the same time, it is not a common case to do this.

withNamespaceConfig :: forall m a b c. (MonadIO m, Ord b) => String -> (TraceConfig -> Namespace a -> m b) -> (Maybe b -> Trace m c -> m (Trace m a)) -> Trace m c -> m (Trace m a)

Take a selector function called extract. Take a function from trace to trace with this config dependent value. In this way construct a trace transformer with a config value

filterSeverityFromConfig :: MonadIO m => Trace m a -> m (Trace m a)

Filter a trace by severity and take the filter value from the config

withDetailsFromConfig :: MonadIO m => Trace m a -> m (Trace m a)

Set detail level of a trace from the config

withBackendsFromConfig :: MonadIO m => (Maybe [BackendConfig] -> Trace m FormattedMessage -> m (Trace m a)) -> m (Trace m a)

Routing and formatting of a trace from the config

withLimitersFromConfig :: forall a m. MonadUnliftIO m => Trace m TraceDispatcherMessage -> Trace m a -> m (Trace m a)

Routing and formatting of a trace from the config

maybeSilent :: forall m a. MonadIO m => (TraceConfig -> Namespace a -> Bool) -> [Text] -> Bool -> Trace m a -> m (Trace m a)

Switch off any message of a particular tracer based on the configuration. If the top tracer is silent and no subtracer is not silent, then switch it off

isSilentTracer :: forall a. MetaTrace a => TraceConfig -> Namespace a -> Bool

hasNoMetrics :: forall a. MetaTrace a => TraceConfig -> Namespace a -> Bool

getSeverity :: TraceConfig -> Namespace a -> SeverityF

By guaranteeing fallback values in the config's namespace root, the Debug severity value here should never get used. If it has to be used, this points to a severely erroneous config, or an implementation error in the application's MetaTrace instance(s). For those cases only, this ensures the affected trace message isn't silently filtered out.

getDetails :: TraceConfig -> Namespace a -> DetailLevel

If no details can be found in the config, it is set to DNormal

getBackends :: TraceConfig -> Namespace a -> [BackendConfig]

If no backends can be found in the config, it is set to [EKGBackend, Forwarder, Stdout HumanFormatColoured]