Hermod.ReCon.Common.Types

Documentation

type NatValue = Word64

type IntValue = Int64

The numeric type used for integer event property values and polynomial arithmetic throughout the framework.

type VariableIdentifier = Text

The type of variable identifiers used throughout the framework (LTL property variables and polynomial integer variables).

type Parser = Parsec Void Text

data BinRel

Binary comparison relation on an ordered type.

Constructors

Eq

Lt

Lte

Gt

Gte