Hermod.ReCon.LTL.Check

Documentation

checkIntTerm :: Set VariableIdentifier -> IntTerm -> [Error]

Validate an integer term by checking all of its free variables are bound.

checkTextTerm :: Set VariableIdentifier -> TextTerm -> [Error]

Validate a text term by checking all of its free variables are bound.

checkParamConstraint :: Set VariableIdentifier -> PropConstraint -> [Error]

Validate a constraint by ensuring its term is well-scoped.

checkFormula :: Set VariableIdentifier -> Formula event ty -> [Error]

Check whether the formula is syntactically valid.

data Error

Constructors

UnboundVariableIdentifier VariableIdentifier

prettyError :: Error -> Text