Hermod.Tracing

Description

Batteries-included public interface for the Hermod tracing system.

A typical application wires up tracing in three steps:

Define trace types: for each domain-specific message type, write LogFormatting (human/machine rendering, metrics) and MetaTrace (namespace, severity, documentation) instances.

Construct backends: call mkHermodTracer (or mkHermodTracer') with standardTracer, ekgTracer, and/or forwardTracer to build a 'Trace IO YourType'.

Configure: load a TraceConfig with readConfiguration and apply it with configureTracers. Use checkTraceConfiguration to validate the config against all known namespaces at startup.

Documentation

class Contravariant (f :: Type -> Type) where

The class of contravariant functors.

Whereas in Haskell, one can think of a Functor as containing or producing values, a contravariant functor is a functor that can be thought of as consuming values.

As an example, consider the type of predicate functions a -> Bool. One such predicate might be negative x = x < 0, which classifies integers as to whether they are negative. However, given this predicate, we can re-use it in other situations, providing we have a way to map values to integers. For instance, we can use the negative predicate on a person's bank balance to work out if they are currently overdrawn:

newtype Predicate a = Predicate { getPredicate :: a -> Bool }

instance Contravariant Predicate where
  contramap :: (a' -> a) -> (Predicate a -> Predicate a')
  contramap f (Predicate p) = Predicate (p . f)
                                         |   `- First, map the input...
                                         `----- then apply the predicate.

overdrawn :: Predicate Person
overdrawn = contramap personBankBalance negative

Any instance should be subject to the following laws:

Identity

contramap id = id

Composition

contramap (g . f) = contramap f . contramap g

Note, that the second law follows from the free theorem of the type of contramap and the first law, so you need only check that the former condition holds.

Minimal complete definition

contramap

Methods

contramap :: (a' -> a) -> f a -> f a'

(>$) :: b -> f b -> f a infixl 4

Replace all locations in the output with the same value. The default definition is contramap . const, but this may be overridden with a more efficient version.

newtype Trace (m :: Type -> Type) a

Constructors

Trace

Fields

unpackTrace :: Tracer m (LoggingContext, Either TraceControl a)

data DataPoint where

Type wrapper for some value of type v. The only reason we need this wrapper is an ability to store different values in the same DataPointStore.

Please note that when the acceptor application will read the value of type v from the store, this value is just as unstructured JSON, but not Haskell value of type v. That's why FromJSON instance for type v should be available for the acceptor application, to decode unstructured JSON.

Constructors

DataPoint :: (ToJSON v, NFData v) => !v -> DataPoint

data ConfigReflection

Constructors

ConfigReflection

Fields

crSilent :: IORef (Set [Text])

crNoMetrics :: IORef (Set [Text])

crAllTracers :: IORef (Set [Text])

data HermodException

Constructors

HermodConfigException

Fields

excMessage :: String

data FormattedMessage

Constructors

FormattedHuman Bool Text
The bool specifies if the formatting includes colours

FormattedMachine Text

FormattedMetrics [Metric]

FormattedForwarder TraceObject

FormattedCBOR ByteString

data ConfigSource

The configuration source that should be ingested

Constructors

FromFile FilePath

FromJSONObject Object

FromLazyBytes ByteString

FromStrictBytes ByteString

type NSWarnings = [Text]

Warnings as a list of text

data PreFormatted

Constructors

PreFormatted

Fields

pfTime :: !UTCTime

pfNamespace :: !Text

pfThreadId :: !Text

pfForHuman :: !(Maybe Text)

pfForMachineObject :: Object

data TraceObject

Used as interface object for ForwarderTracer

Constructors

TraceObject

Fields

toHuman :: !(Maybe Text)

toMachine :: !Text

toNamespace :: ![Text]

toSeverity :: !SeverityS

toDetails :: !DetailLevel

toTimestamp :: !UTCTime

toHostname :: !Text

toThreadId :: !Text

data LimiterSpec

Constructors

LimiterSpec

Fields

lsNs :: [Text]

lsName :: Text

lsFrequency :: Double

type DataPointName = Text

type DataPointStore = TVar (Map DataPointName DataPoint)

data HowToConnect

Constructors

LocalPipe !FilePath
Local pipe (UNIX or Windows).

RemoteSocket !Host !Port
Remote socket (host and port).

type Host = Text

Specifies how to connect to the peer.

Taken from ekg-forward:System.Metrics.Configuration, to avoid dependency.

type Port = Word16

data SeverityS

Constructors

Debug

Info

Notice

Warning

Error

Critical

Alert

Emergency

data Namespace a

Constructors

Namespace

Fields

nsPrefix :: [Text]

nsInner :: [Text]

data TraceOptionForwarder

Constructors

TraceOptionForwarder

Fields

tofQueueSize :: Word

tofVerbosity :: Verbosity

tofMaxReconnectDelay :: Word

data PrometheusSimpleRun

Constructors

PrometheusSimpleRun

Fields

connTimeout :: Maybe Word

connCountGlobal :: Maybe Word

connCountPerHost :: Maybe Word

connPerSecond :: Maybe Double

newtype SeverityF

Constructors

SeverityF (Maybe SeverityS)

data DetailLevel

Constructors

DMinimal

DNormal

DDetailed

DMaximum

data BackendConfig

Constructors

Forwarder

Stdout FormatLogging

EKGBackend

DatapointBackend

PrometheusSimple Bool (Maybe HostName) PortNumber

data TraceConfig

Constructors

TraceConfig

Fields

tcOptions :: Map [Text] [ConfigOption]

tcForwarder :: Maybe TraceOptionForwarder

tcApplicationName :: Maybe Text

tcMetricsPrefix :: Maybe Text

tcPeriodicTracers :: Map Text Word64

tcPrometheusSimpleRun :: Maybe PrometheusSimpleRun

data ConfigOption

Constructors

ConfSeverity

Fields

severity :: SeverityF

ConfDetail

Fields

detail :: DetailLevel

ConfBackend

Fields

backends :: [BackendConfig]

ConfLimiter

Fields

maxFrequency :: Double

data CounterAction

Constructors

CounterIncrement

CounterAdd Word64

data LoggingContext

Constructors

LoggingContext

Fields

lcNSInner :: [Text]

lcNSPrefix :: [Text]

lcSeverity :: Maybe SeverityS

lcPrivacy :: Maybe Privacy

lcDetails :: Maybe DetailLevel

data Metric

Constructors

IntM Text Integer

DoubleM Text Double

CounterM Text CounterAction

LabelSetM Text [(Text, Text)]

data Privacy

Constructors

Confidential

Public

data FormatLogging

Constructors

HumanFormatColoured

HumanFormatUncoloured

MachineFormat

data ForwarderMode

Constructors

Initiator

Responder

data Verbosity

Constructors

Maximum

Minimum

newtype DocCollector

Constructors

DocCollector (IORef (Map Int LogDoc))

data LogDoc

Constructors

LogDoc

Fields

ldDoc :: !Text

ldMetricsDoc :: !(Map Text Text)

ldNamespace :: ![([Text], [Text])]

ldSeverityCoded :: !(Maybe SeverityS)

ldPrivacyCoded :: !(Maybe Privacy)

ldDetailsCoded :: !(Maybe DetailLevel)

ldDetails :: ![DetailLevel]

ldBackends :: ![BackendConfig]

ldFiltered :: ![SeverityF]

ldLimiter :: ![(Text, Double)]

ldSilent :: Bool

class LogFormatting a where

Minimal complete definition

forMachine

Methods

forMachine :: DetailLevel -> a -> Object

forHuman :: a -> Text

asMetrics :: a -> [Metric]

newtype Folding a b

Constructors

Folding b

class MetaTrace a where

Minimal complete definition

namespaceFor, severityFor, documentFor, allNamespaces

Methods

namespaceFor :: a -> Namespace a

severityFor :: Namespace a -> Maybe a -> Maybe SeverityS

privacyFor :: Namespace a -> Maybe a -> Maybe Privacy

detailsFor :: Namespace a -> Maybe a -> Maybe DetailLevel

documentFor :: Namespace a -> Maybe Text

metricsDocFor :: Namespace a -> [(Text, Text)]

allNamespaces :: [Namespace a]

data TraceControl where

Constructors

TCReset :: TraceControl

TCConfig :: TraceConfig -> TraceControl

TCOptimize :: ConfigReflection -> TraceControl

TCDocument :: Int -> DocCollector -> TraceControl

traceWith :: Monad m => Trace m a -> a -> m ()

unfold :: Folding a b -> b

(>$<) :: Contravariant f => (a -> b) -> f b -> f a infixl 4

This is an infix alias for contramap.

arrow :: forall (m :: Type -> Type) a. TracerA m a () -> Tracer m a

contramapM :: Monad m => Trace m b -> ((LoggingContext, Either TraceControl a) -> m (LoggingContext, Either TraceControl b)) -> Trace m a

emptyConfigReflection :: IO ConfigReflection

configureTracers :: forall a m. (MetaTrace a, MonadIO m) => ConfigReflection -> TraceConfig -> [Trace m a] -> m ()

Call this function at initialisation, and later for reconfiguration. Config reflection is used to optimise the tracers and has to collect information about the tracers. Although it is possible to give more then one tracer of the same time, it is not a common case to do this.

withNamespaceConfig :: forall m a b c. (MonadIO m, Ord b) => String -> (TraceConfig -> Namespace a -> m b) -> (Maybe b -> Trace m c -> Trace m a) -> Trace m c -> m (Trace m a)

Take a selector function called extract. Take a function from trace to trace with this config dependent value. In this way construct a trace transformer with a config value

filterSeverityFromConfig :: MonadIO m => Trace m a -> m (Trace m a)

Filter a trace by severity and take the filter value from the config

withDetailsFromConfig :: MonadIO m => Trace m a -> m (Trace m a)

Set detail level of a trace from the config

withBackendsFromConfig :: MonadIO m => (Maybe [BackendConfig] -> Trace m FormattedMessage -> Trace m a) -> m (Trace m a)

Routing and formatting of a trace from the config

withLimitersFromConfig :: forall a m. MonadUnliftIO m => Trace m HermodTracingMessage -> Trace m a -> m (Trace m a)

Routing and formatting of a trace from the config

maybeSilent :: forall m a. MonadIO m => (TraceConfig -> Namespace a -> Bool) -> [Text] -> Bool -> Trace m a -> m (Trace m a)

Switch off any message of a particular tracer based on the configuration. If the top tracer is silent and no subtracer is not silent, then switch it off

isSilentTracer :: forall a. MetaTrace a => TraceConfig -> Namespace a -> Bool

hasNoMetrics :: forall a. MetaTrace a => TraceConfig -> Namespace a -> Bool

getSeverity :: TraceConfig -> Namespace a -> SeverityF

By guaranteeing fallback values in the config's namespace root, the Debug severity value here should never get used. If it has to be used, this points to a severely erroneous config, or an implementation error in the application's MetaTrace instance(s). For those cases only, this ensures the affected trace message isn't silently filtered out.

getDetails :: TraceConfig -> Namespace a -> DetailLevel

If no details can be found in the config, it is set to DNormal

getBackends :: TraceConfig -> Namespace a -> [BackendConfig]

If no backends can be found in the config, it is set to [EKGBackend, Forwarder, Stdout HumanFormatColoured]

limitFrequency :: forall a m. MonadUnliftIO m => Double -> Text -> Trace m HermodTracingMessage -> Trace m a -> m (Trace m a)

Limits the frequency of messages to nMsg which is given per minute.

If the limiter detects more messages, it traces randomly selected messages with the given frequency on the vtracer until the frequency falls under the threshold long enough.(see below)

Before this the ltracer gets a StartLimiting message. In-between you receive ContinueLimiting messages on the ltracer every reminderPeriod seconds, with the number of suppressed messages. Finally it sends a StopLimiting message on the ltracer and traces all messages on the vtracer again.

A budget is used to decide when to start limiting and stop limiting, so that the limiter does not get activated if few messages are send in high frequency, and doesn't get deactivated if their are only few messages which come with low frequency. When messages arrive in shorter frequency then by the given thresholdFrequency budget is earned, and if they arrive in a longer period budget is spend. If budget is gets higher then budgetLimit, the limiter starts, and if it falls below minus budgetLimit the limiter stops.

mkConfiguration :: TraceConfig

Creates the minimal viable configuration by only setting fallback values in an empty TraceConfig. Fallback options for the namespace root: Notice severity, normal detail, JSON stdout logging. Notice severity was chosen as it will never filter out any actionable traces while creating minimal noise in the log.

mkConfigurationWithFallback :: SeverityS -> DetailLevel -> BackendConfig -> TraceConfig

Creates the minimal viable configuration by only setting custom fallback values in an empty TraceConfig. Fallback options for the namespace root: custom values.

readConfiguration :: ConfigSource -> IO TraceConfig

Read a configuration source and return the internal representation. Fallback options for the namespace root: Notice severity, normal detail, JSON stdout logging.

readConfigurationWithFallback :: SeverityS -> DetailLevel -> BackendConfig -> ConfigSource -> IO TraceConfig

Read a configuration source and return the internal representation. Fallback options for the namespace root: custom values.

readConfiguration' :: FilePath -> IO TraceConfig

Read a configuration file and return the internal representation. This will silently provide a minimal viable config via mkConfiguration when the file is absent. Fallback options for the namespace root: Notice severity, normal detail, JSON stdout logging.

readConfigurationWithFallback' :: SeverityS -> DetailLevel -> BackendConfig -> FilePath -> IO TraceConfig

Read a configuration file and return the internal representation. This will silently provide a minimal viable config via mkConfigurationWithFallback when the file is absent. Fallback options for the namespace root: custom values.

readConfigurationWithDefault :: ConfigSource -> TraceConfig -> IO TraceConfig

Read a configuration source and return the internal representation. TraceConfig fields not specified in the file will be taken from the provided defaultConf (when given there). Fallback options for the namespace root: Notice severity, normal detail, JSON stdout logging.

readConfigurationWithFallbackAndDefault :: SeverityS -> DetailLevel -> BackendConfig -> ConfigSource -> TraceConfig -> IO TraceConfig

Read a configuration source and return the internal representation. TraceConfig fields not specified in the file will be taken from the provided defaultConf (when given there). Fallback options for the namespace root: custom values.

applyFallback :: SeverityS -> DetailLevel -> BackendConfig -> TraceConfig -> TraceConfig

Applies the fallback values to the namespace root, or creates a namespace root from them if none is present. Furthermore, it ensures a metric prefix is properly namespaced, if there is one configured. If you do not use any of mkConfiguration* or readConfiguration* to create your TraceConfig, but do it manually, it is highly recommended to call applyFallback on that TraceConfig value as a last step before using it.

configToRepresentation :: TraceConfig -> ConfigRepresentation

Convert config from internal to external representation

checkTraceConfiguration :: ConfigSource -> TraceConfig -> [([Text], [Text])] -> IO NSWarnings

Checks if all namespaces in this configuration are legal. Legal in this case means that it can be found by a hierarchical lookup in all namespaces. Warns if namespaces in all namespaces are not unique, Warns if namespaces in all namespaces are ending in the middle of another namespace. The namespaces in allNamespaces are consistent with the namespaces for the severityFor, privacyFor, detailsFor, documentFor and metricsDocFor functions.

checkTraceConfiguration' :: TraceConfig -> [([Text], [Text])] -> NSWarnings

metricsFormatter :: forall a m. (LogFormatting a, MonadIO m) => Trace m FormattedMessage -> Trace m a

Format this trace as metrics

preFormatted :: (LogFormatting a, MonadIO m) => Bool -> Trace m PreFormatted -> Trace m a

Transform this trace to a preformatted message, so that double serialization is avoided

forwardFormatter :: forall a m. (MonadIO m, LogFormatting a) => Trace m FormattedMessage -> Trace m a

forwardFormatter' :: forall m. MonadIO m => Trace m FormattedMessage -> Trace m PreFormatted

Format this trace as TraceObject for the trace forwarder

machineFormatter :: forall a m. (MonadIO m, LogFormatting a) => Trace m FormattedMessage -> Trace m a

machineFormatter' :: forall m. MonadIO m => Trace m FormattedMessage -> Trace m PreFormatted

Format this trace as TraceObject for machine-readable text output (JSON)

cborFormatter :: forall a m. (MonadIO m, LogFormatting a) => Trace m FormattedMessage -> Trace m a

cborFormatter' :: forall m. MonadIO m => Trace m FormattedMessage -> Trace m PreFormatted

Format this trace in binary serialisation (CBOR)

humanFormatter :: forall a m. (MonadIO m, LogFormatting a) => Bool -> Trace m FormattedMessage -> Trace m a

humanFormatter' :: forall m. MonadIO m => Bool -> Trace m FormattedMessage -> Trace m PreFormatted

Format this trace in human readable text output

showT :: Show a => a -> Text

Convenience function for a Show instance to be converted to text immediately

filterTrace :: forall (m :: Type -> Type) a. Monad m => ((LoggingContext, a) -> Bool) -> Trace m a -> Trace m a

initDataPointStore :: IO DataPointStore

writeToStore :: DataPointStore -> DataPointName -> DataPoint -> IO ()

Write DataPoint to the store.

dataPointTracer :: forall m. MonadIO m => DataPointStore -> Trace m DataPoint

mkDataPointTracer :: forall dp. (ToJSON dp, MetaTrace dp, NFData dp) => Trace IO DataPoint -> IO (Trace IO dp)

tryEvalNF :: NFData a => a -> IO (Either SomeException a)

ekgTracer :: MonadIO m => TraceConfig -> Store -> m (Trace m FormattedMessage)

It is mandatory to construct only one standard tracer in any application! Throwing away a standard tracer and using a new one will result in an exception

showTReal :: RealFloat a => a -> Text

standardTracer :: forall m. MonadIO m => m (Trace m FormattedMessage)

The standardTracer handles stdout logging in a thread-safe manner. It is strongly advised to construct only one standardTracer for any application.

threadLabelMe :: String -> IO ()

forwardTracer :: forall m. MonadIO m => (TraceObject -> IO ()) -> Trace m FormattedMessage

It is mandatory to construct only one forwardTracer tracer in any application! Throwing away a forwardTracer tracer and using a new one will result in an exception

mkHermodTracer :: forall evt. (LogFormatting evt, MetaTrace evt) => Trace IO FormattedMessage -> Trace IO FormattedMessage -> Maybe (Trace IO FormattedMessage) -> [Text] -> IO (Trace IO evt)

Construct a hermod tracer. The tracer gets a name, which is appended to its namespace. The tracer has to be an instance of LogFormatting for the display of messages and an instance of MetaTrace for meta information such as severity, privacy, details and backends'. The tracer gets the backends': trStdout, trForward and mbTrEkg as arguments. The returned tracer needs to be configured with a configuration before it is used.

mkHermodTracer' :: forall evt evt1. (LogFormatting evt1, MetaTrace evt1) => Trace IO FormattedMessage -> Trace IO FormattedMessage -> Maybe (Trace IO FormattedMessage) -> [Text] -> (Trace IO evt1 -> IO (Trace IO evt)) -> IO (Trace IO evt)

Adds the possibility to add special tracers via the hook function

mkMetricsTracer :: Maybe (Trace IO FormattedMessage) -> Trace IO FormattedMessage

traceTracerInfo :: Trace IO FormattedMessage -> Trace IO FormattedMessage -> ConfigReflection -> IO ()

traceConfigWarnings :: Trace IO FormattedMessage -> Trace IO FormattedMessage -> [Text] -> IO ()

traceEffectiveConfiguration :: Trace IO FormattedMessage -> Trace IO FormattedMessage -> TraceConfig -> IO ()

runInLoop :: IO () -> (SomeException -> IO ()) -> Word64 -> Word64 -> IO ()

Run an IO action which may throw an exception in a loop. On exception, the action will be re-run after a pause (the delay parameters represent seconds). That pause doubles which each exception, but is reset when the action runs long enough.

emptyTraceConfig :: TraceConfig

emptyLoggingContext :: LoggingContext

getMetricName :: Metric -> Text

nsCast :: Namespace a -> Namespace b

nsGetComplete :: Namespace a -> [Text]

nsGetTuple :: Namespace a -> ([Text], [Text])

nsPrependInner :: Text -> Namespace a -> Namespace b

nsRawToText :: ([Text], [Text]) -> Text

nsReplaceInner :: [Text] -> Namespace a -> Namespace a

nsReplacePrefix :: [Text] -> Namespace a -> Namespace a

nsToText :: Namespace a -> Text

defaultForwarder :: TraceOptionForwarder

parsePrometheusString :: Text -> Either String BackendConfig

prometheusSimpleNoOverrides :: PrometheusSimpleRun

emptyLogDoc :: Text -> [(Text, Text)] -> LogDoc

contramapMCond :: Monad m => Trace m b -> ((LoggingContext, Either TraceControl a) -> m (Maybe (LoggingContext, Either TraceControl b))) -> Trace m a

foldCondTraceM :: forall a acc m. MonadUnliftIO m => (acc -> LoggingContext -> a -> m acc) -> acc -> (a -> Bool) -> Trace m (Folding a acc) -> m (Trace m a)

foldTraceM :: forall a acc m. MonadUnliftIO m => (acc -> LoggingContext -> a -> m acc) -> acc -> Trace m (Folding a acc) -> m (Trace m a)

routingTrace :: Monad m => (a -> m (Trace m a)) -> Trace m a -> Trace m a

filterTraceMaybe :: forall (m :: Type -> Type) a. Monad m => Trace m a -> Trace m (Maybe a)

effect :: (a -> m b) -> TracerA m a b

emit :: Applicative m => (a -> m ()) -> TracerA m a ()

nat :: (forall x. m x -> n x) -> TracerA m a b -> TracerA n a b

squelch :: forall (m :: Type -> Type) a. Applicative m => TracerA m a ()

runTracer :: Tracer m a -> TracerA m a ()

debugTracer :: forall (m :: Type -> Type). Applicative m => Tracer m String

mkTracer :: Applicative m => (a -> m ()) -> Tracer m a

natTracer :: (forall x. m x -> n x) -> Tracer m s -> Tracer n s

squelchUnless :: forall (m :: Type -> Type) a. Monad m => (a -> Bool) -> Tracer m a -> Tracer m a

squelchUnlessM :: Monad m => (a -> m Bool) -> Tracer m a -> Tracer m a

stdoutTracer :: Tracer IO String

traceAll :: forall (m :: Type -> Type) t b a. (Monad m, Foldable t) => (b -> t a) -> Tracer m a -> Tracer m b

traceMaybe :: forall (m :: Type -> Type) a b. Monad m => (a -> Maybe b) -> Tracer m b -> Tracer m a

traceMaybeM :: Monad m => (a -> m (Maybe b)) -> Tracer m b -> Tracer m a

traceTraversable :: forall (m :: Type -> Type) t a. (Monad m, Foldable t) => Tracer m a -> Tracer m (t a)

use :: forall (m :: Type -> Type) a. Tracer m a -> TracerA m a ()

excludeAsyncExceptions :: SomeException -> Maybe SomeException

Helper functions to force trace + metric values in a controlled section of code

showTHex :: Integral a => a -> Text