Hermod.Tracing.API

Description

Stable public API for the Hermod tracing system.

This is the single-import front door for hermod-tracing-api. It re-exports everything a package needs to:

Define trace types: write LogFormatting (human/machine rendering, metrics) and MetaTrace (namespace, severity, documentation) instances for your domain message types.

Dispatch messages: call traceWith to emit, contramapM to adapt types, foldTraceM to accumulate state, routingTrace to fan out.

Filter: filterTrace, filterTraceMaybe.

For tracer authors

Trace                  -- the central carrier opaque type
LogFormatting(..)      -- typeclass: forMachine, forHuman, asMetrics
MetaTrace(..)          -- typeclass: namespaceFor, severityFor, documentFor, …
Metric(..)             -- metric payload (IntM, DoubleM, CounterM, LabelSetM)
Namespace(..)          -- hierarchical trace identifier
SeverityS(..)          -- message severity (Debug … Emergency)
SeverityF(..)          -- severity filter (Nothing = Silence)
Privacy(..)            -- Public | Confidential
DetailLevel(..)        -- DMinimal … DMaximum
Folding(..)            -- wrapper for fold-based stateful tracers

Configuration and control (consumed by hermod-tracing-core)

TraceConfig, ConfigOption, BackendConfig, ConfigReflection, DocCollector, ForwarderAddr, ForwarderMode, TraceOptionForwarder, PrometheusSimpleRun. These appear in type signatures throughout the system; tracer authors typically do not construct them directly.

Documentation

data Trace (m :: Type -> Type) a

newtype Folding a b

Constructors

Folding b

data CounterAction

Constructors

CounterIncrement

CounterAdd Word64

data DetailLevel

Constructors

DMinimal

DNormal

DDetailed

DMaximum

data Metric

Constructors

IntM Text Integer

DoubleM Text Double

CounterM Text CounterAction

LabelSetM Text [(Text, Text)]

data Namespace a

Constructors

Namespace

Fields

nsPrefix :: [Text]

nsInner :: [Text]

data Privacy

Constructors

Confidential

Public

newtype SeverityF

Constructors

SeverityF (Maybe SeverityS)

data SeverityS

Constructors

Debug

Info

Notice

Warning

Error

Critical

Alert

Emergency

data BackendConfig

Constructors

Forwarder

Stdout FormatLogging

EKGBackend

DatapointBackend

PrometheusSimple Bool (Maybe HostName) PortNumber

data ConfigOption

Constructors

ConfSeverity

Fields

severity :: SeverityF

ConfDetail

Fields

detail :: DetailLevel

ConfBackend

Fields

backends :: [BackendConfig]

ConfLimiter

Fields

maxFrequency :: Double

data FormatLogging

Constructors

HumanFormatColoured

HumanFormatUncoloured

MachineFormat

data ForwarderMode

Constructors

Initiator

Responder

data PrometheusSimpleRun

Constructors

PrometheusSimpleRun

Fields

connTimeout :: Maybe Word

connCountGlobal :: Maybe Word

connCountPerHost :: Maybe Word

connPerSecond :: Maybe Double

data TraceConfig

Constructors

TraceConfig

Fields

tcOptions :: Map [Text] [ConfigOption]

tcForwarder :: Maybe TraceOptionForwarder

tcApplicationName :: Maybe Text

tcMetricsPrefix :: Maybe Text

tcPeriodicTracers :: Map Text Word64

tcPrometheusSimpleRun :: Maybe PrometheusSimpleRun

data TraceOptionForwarder

Constructors

TraceOptionForwarder

Fields

tofQueueSize :: Word

tofVerbosity :: Verbosity

tofMaxReconnectDelay :: Word

data Verbosity

Constructors

Maximum

Minimum

data ConfigReflection

Constructors

ConfigReflection

Fields

crSilent :: IORef (Set [Text])

crNoMetrics :: IORef (Set [Text])

crAllTracers :: IORef (Set [Text])

newtype DocCollector

Constructors

DocCollector (IORef (Map Int LogDoc))

class LogFormatting a where

Minimal complete definition

forMachine

Methods

forMachine :: DetailLevel -> a -> Object

forHuman :: a -> Text

asMetrics :: a -> [Metric]

class MetaTrace a where

Minimal complete definition

namespaceFor, severityFor, documentFor, allNamespaces

Methods

namespaceFor :: a -> Namespace a

severityFor :: Namespace a -> Maybe a -> Maybe SeverityS

privacyFor :: Namespace a -> Maybe a -> Maybe Privacy

detailsFor :: Namespace a -> Maybe a -> Maybe DetailLevel

documentFor :: Namespace a -> Maybe Text

metricsDocFor :: Namespace a -> [(Text, Text)]

allNamespaces :: [Namespace a]

traceWith :: Monad m => Trace m a -> a -> m ()

unfold :: Folding a b -> b

routingTrace :: Monad m => (a -> m (Trace m a)) -> Trace m a -> Trace m a

filterTraceMaybe :: forall (m :: Type -> Type) a. Monad m => Trace m a -> Trace m (Maybe a)

emptyLoggingContext :: LoggingContext

getMetricName :: Metric -> Text

nsCast :: Namespace a -> Namespace b

nsGetComplete :: Namespace a -> [Text]

nsGetTuple :: Namespace a -> ([Text], [Text])

nsPrependInner :: Text -> Namespace a -> Namespace b

nsRawToText :: ([Text], [Text]) -> Text

nsReplaceInner :: [Text] -> Namespace a -> Namespace a

nsReplacePrefix :: [Text] -> Namespace a -> Namespace a

nsToText :: Namespace a -> Text

defaultForwarder :: TraceOptionForwarder

emptyTraceConfig :: TraceConfig

parsePrometheusString :: Text -> Either String BackendConfig

prometheusSimpleNoOverrides :: PrometheusSimpleRun

emptyConfigReflection :: IO ConfigReflection

emptyLogDoc :: Text -> [(Text, Text)] -> LogDoc

contramapM :: Monad m => Trace m b -> (a -> m b) -> Trace m a

Contramap a monadic function over a trace.

contramapMCond :: Monad m => Trace m b -> (a -> m (Maybe b)) -> Trace m a

Like contramapM but can also filter out messages by returning Nothing.

foldTraceM :: forall a acc m. MonadUnliftIO m => (acc -> a -> m acc) -> acc -> Trace m (Folding a acc) -> m (Trace m a)

Fold a monadic accumulator function over a trace. Uses an MVar to hold the state.

foldCondTraceM :: forall a acc m. MonadUnliftIO m => (acc -> a -> m acc) -> acc -> (a -> Bool) -> Trace m (Folding a acc) -> m (Trace m a)

Like foldTraceM but additionally filter the trace by a predicate.

filterTrace :: Monad m => (a -> Bool) -> Trace m a -> Trace m a