Hermod.Tracing.Consistency

Documentation

type NSWarnings = [Text]

Warnings as a list of text

checkTraceConfiguration :: ConfigSource -> TraceConfig -> [([Text], [Text])] -> IO NSWarnings

Checks if all namespaces in this configuration are legal. Legal in this case means that it can be found by a hierarchical lookup in all namespaces. Warns if namespaces in all namespaces are not unique, Warns if namespaces in all namespaces are ending in the middle of another namespace. The namespaces in allNamespaces are consistent with the namespaces for the severityFor, privacyFor, detailsFor, documentFor and metricsDocFor functions.

checkTraceConfiguration' :: TraceConfig -> [([Text], [Text])] -> NSWarnings