Hermod.Tracing.HermodTracingMessage

Documentation

data UnknownNamespaceKind

Constructors

UKFSeverity

UKFPrivacy

UKFDetails

data HermodTracingMessage

Constructors

StartLimiting Text
This message indicates the start of frequency limiting

StopLimiting Text Int
This message indicates the stop of frequency limiting, and gives the number of messages that has been suppressed

RememberLimiting Text Int
This message remembers of ongoing frequency limiting, and gives the number of messages that has been suppressed

UnknownNamespace [Text] [Text] UnknownNamespaceKind
An internal error was detected

TracerInfo [Text] [Text] [Text]
The first array signifies the namespace of silent tracers The second array signifies the namespace tracers without metrics The third array gives the names of all tracers

MetricsInfo (Map Text Int)
Outputs optional statistics about metrics frequency

TracerConsistencyWarnings [Text]
Consistency check found warnings

TracerInfoConfig TraceConfig
Trace the effective configuration as JSON