Hermod.Tracing.Tracer.Composed

Documentation

mkHermodTracer :: forall evt. (LogFormatting evt, MetaTrace evt) => Trace IO FormattedMessage -> Trace IO FormattedMessage -> Maybe (Trace IO FormattedMessage) -> [Text] -> IO (Trace IO evt)

Construct a hermod tracer. The tracer gets a name, which is appended to its namespace. The tracer has to be an instance of LogFormatting for the display of messages and an instance of MetaTrace for meta information such as severity, privacy, details and backends'. The tracer gets the backends': trStdout, trForward and mbTrEkg as arguments. The returned tracer needs to be configured with a configuration before it is used.

mkHermodTracer' :: forall evt evt1. (LogFormatting evt1, MetaTrace evt1) => Trace IO FormattedMessage -> Trace IO FormattedMessage -> Maybe (Trace IO FormattedMessage) -> [Text] -> (Trace IO evt1 -> IO (Trace IO evt)) -> IO (Trace IO evt)

Adds the possibility to add special tracers via the hook function

mkMetricsTracer :: Maybe (Trace IO FormattedMessage) -> Trace IO FormattedMessage

traceTracerInfo :: Trace IO FormattedMessage -> Trace IO FormattedMessage -> ConfigReflection -> IO ()

traceConfigWarnings :: Trace IO FormattedMessage -> Trace IO FormattedMessage -> [Text] -> IO ()

traceEffectiveConfiguration :: Trace IO FormattedMessage -> Trace IO FormattedMessage -> TraceConfig -> IO ()